feat: update rand and stop relying on thread_rng
Summary:
Now we can have the user pass in their own rng. This also allows
deterministic usage.

While updating. I added the rps feature. It is not on by deault. Since
that also is the only place the once_cell is used, we drop that
dependecy by default.

Test Plan:
cargo test